Split TextField
