+++ src/app/components/molecules/Input/Input.sc.ts
@@ -1,10 +1,10 @@
-import { AdornmentPosition, InputType, InputVariant } from '../../../types';
+import { AdornmentPosition, InputVariant } from '../../../types';
 import styled, { css, SimpleInterpolation } from 'styled-components';
 import { getBorderColor } from '../../../styles/mixins/getBorderColor';
 import { setBoxSizing } from '../../../styles/mixins/setBoxSizing';
 import { themeBasic } from '../../../styles/theming/themes/basic';
 
-interface StyledInputBaseProps {
+export interface StyledInputBaseProps {
     hasError: boolean;
     isDisabled: boolean;
     isFocused: boolean;
@@ -70,87 +70,6 @@ export const StyledInput = styled.div<StyledInputProps>`
 `;
 
 StyledInput.defaultProps = {
-    theme: themeBasic,
-};
-
-interface TextFieldProps extends StyledInputBaseProps {
-    adornmentPosition: AdornmentPosition;
-    hasAdornment: boolean;
-    hasNegativeAmountColor: boolean;
-    isTextarea: boolean;
-    type: InputType;
-}
-
-export const TextField = styled.input<TextFieldProps>`
-    ${({ theme }): string => theme.textStyling(theme.availableTextStyles().body1)}
-    display: block;
-    outline: none;
-    background-color: transparent;
-    width: 100%;
-    color: ${({ theme }): string => theme.colorTextBody.primary};
-
-    ${({ hasNegativeAmountColor, theme }): SimpleInterpolation =>
-        hasNegativeAmountColor &&
-        css`
-            color: ${theme.colorInvalid};
-        `}
-
-    ${({ adornmentPosition, hasAdornment, theme }): SimpleInterpolation =>
-        hasAdornment &&
-        adornmentPosition === AdornmentPosition.LEFT &&
-        css`
-            text-indent: ${theme.spacing(2)};
-        `}
-
-    ${({ theme, variant }): SimpleInterpolation =>
-        variant === InputVariant.COMPACT &&
-        css`
-            border: 0;
-            border-bottom: 1px solid;
-            padding: 0;
-            height: ${theme.spacing(3.5)};
-        `}
-
-    ${({ theme, variant }): SimpleInterpolation =>
-        variant === InputVariant.OUTLINE &&
-        css`
-            border: 1px solid;
-            border-radius: ${theme.spacing(1)};
-            padding: ${theme.spacing(0, 1.5)};
-            height: ${theme.spacing(6)};
-        `}
-
-    ${({ isTextarea, theme }): SimpleInterpolation =>
-        isTextarea &&
-        css`
-            padding: ${theme.spacing(1.5)};
-            height: ${theme.spacing(16)};
-            resize: none;
-        `}
-
-    ${({ isDisabled, theme }): SimpleInterpolation =>
-        isDisabled &&
-        css`
-            color: ${theme.colorDisabled};
-        `}
-
-    ${({ hasError, isDisabled, isFocused, isHovered, isValid, theme }): SimpleInterpolation =>
-        css`
-            /* stylelint-disable indentation */
-            border-color: ${getBorderColor({
-                defaultColor: theme.colorPrimary,
-                hasError,
-                isDisabled,
-                isFocused,
-                isHovered,
-                isValid,
-                theme,
-            })};
-            /* stylelint-enable indentation */
-        `}
-`;
-
-TextField.defaultProps = {
     theme: themeBasic,
 };
